+++ libpiper/src/piper.cpp
@@ -198,14 +198,20 @@ int piper_synthesize_start(struct piper_synthesizer *synth, const char *text,
     }
 
     // phonemes to ids
+    std::vector<Phoneme> sentence_codepoints;
     std::vector<PhonemeId> sentence_ids;
     for (auto &phonemes_str : sentence_phonemes) {
         if (phonemes_str.empty()) {
             continue;
         }
 
+        sentence_codepoints.push_back(PHONEME_BOS);
         sentence_ids.push_back(ID_BOS);
+
+        sentence_codepoints.push_back(PHONEME_BOS);
         sentence_ids.push_back(ID_PAD);
+
+        sentence_codepoints.push_back(PHONEME_SEPARATOR);
 
         auto phonemes_norm = una::norm::to_nfd_utf8(phonemes_str);
         auto phonemes_range = una::ranges::utf8_view{phonemes_norm};
@@ -216,22 +222,28 @@ int piper_synthesize_start(struct piper_synthesizer *synth, const char *text,
         // These surround words from languages other than the current voice.
         bool in_lang_flag = false;
         while (phonemes_iter != phonemes_end) {
+            auto phoneme = *phonemes_iter;
+
             if (in_lang_flag) {
-                if (*phonemes_iter == U')') {
+                if (phoneme == U')') {
                     // End of (lang) switch
                     in_lang_flag = false;
                 }
-            } else if (*phonemes_iter == U'(') {
+            } else if (phoneme == U'(') {
                 // Start of (lang) switch
                 in_lang_flag = true;
             } else {
                 // Look up ids
-                auto ids_for_phoneme =
-                    synth->phoneme_id_map.find(*phonemes_iter);
+                auto ids_for_phoneme = synth->phoneme_id_map.find(phoneme);
                 if (ids_for_phoneme != synth->phoneme_id_map.end()) {
                     for (auto id : ids_for_phoneme->second) {
+                        sentence_codepoints.push_back(phoneme);
                         sentence_ids.push_back(id);
+
+                        sentence_codepoints.push_back(phoneme);
                         sentence_ids.push_back(ID_PAD);
+
+                        sentence_codepoints.push_back(PHONEME_SEPARATOR);
                     }
                 }
             }
@@ -239,9 +251,12 @@ int piper_synthesize_start(struct piper_synthesizer *synth, const char *text,
             phonemes_iter++;
         }
 
+        sentence_codepoints.push_back(PHONEME_EOS);
         sentence_ids.push_back(ID_EOS);
+        sentence_codepoints.push_back(PHONEME_SEPARATOR);
 
-        synth->phoneme_id_queue.emplace(std::move(sentence_ids));
+        synth->phoneme_id_queue.emplace(
+            std::move(std::make_pair(sentence_codepoints, sentence_ids)));
         sentence_ids.clear();
     }
 
@@ -260,7 +275,7 @@ int piper_synthesize_next(struct piper_synthesizer *synth,
 
     // Clear data from previous call
     synth->chunk_samples.clear();
-    synth->chunk_phonemes = "";
+    synth->chunk_phonemes.clear();
     synth->chunk_phoneme_ids.clear();
     synth->chunk_alignments.clear();
 
@@ -280,7 +295,7 @@ int piper_synthesize_next(struct piper_synthesizer *synth,
     }
 
     // Process next list of phoneme ids
-    auto next_ids = std::move(synth->phoneme_id_queue.front());
+    auto [next_phonemes, next_ids] = std::move(synth->phoneme_id_queue.front());
     synth->phoneme_id_queue.pop();
 
     auto memoryInfo = Ort::MemoryInfo::CreateCpu(
@@ -353,6 +368,11 @@ int piper_synthesize_next(struct piper_synthesizer *synth,
     chunk->samples = synth->chunk_samples.data();
 
     chunk->is_last = synth->phoneme_id_queue.empty();
+
+    // Copy phonemes
+    synth->chunk_phonemes = std::move(next_phonemes);
+    chunk->phonemes = synth->chunk_phonemes.data();
+    chunk->num_phonemes = synth->chunk_phonemes.size();
 
     // Copy phoneme ids
     for (auto phoneme_id : next_ids) {